- Warning fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@35249 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csplugincommon/rendermanager/occluvis.cpp
@@ -350,7 +350,7 @@ namespace CS
 
       if (queryData->eResult == VISIBLE)
       {
-        queryData->uNextCheck += visibilityFrameSkip * nodeMeshHash.GetSize ();
+        queryData->uNextCheck += visibilityFrameSkip * (uint32)nodeMeshHash.GetSize ();
       }
 
       return queryData->eResult;
